add part first visit loading
