+++ src/sass/utils/_breakpoints.scss
@@ -1,4 +1,4 @@
-// $mobile:
+// $mobile: 
 // $mobile-max:
 // $tablet:
 // $tablet-max:
