+++ src/sass/utils/_evetns-list.scss
@@ -1,15 +1,29 @@
 @import './utils/placeholders';
 
+.events {
+    background-color: $main-body-color;
 
+}
 
 .events-gallery {
     display: flex;
     flex-wrap: wrap;
-    justify-content: space-between;
+    justify-content: space-around;
     margin-right: 25px;
     margin-left: 25px;
 
-    background-color: black;
+
+
+    @media (min-width: 768px) {
+        justify-content: space-around;
+        margin-right: auto;
+        margin-left: auto;
+        width: 610px;
+    }
+
+    @media (min-width: 1280px) {
+        width: 946px;
+    }
 }
 
 .event-card {
@@ -18,6 +32,18 @@
     min-height: 231px;
     margin-bottom: 33px;
     position: relative;
+
+
+
+    @media (min-width: 768px) {
+        width: 194px;
+        min-height: 312px;
+    }
+
+    @media (min-width: 1280px) {
+        width: 196px;
+        min-height: 312px;
+    }
 }
 
 .rose-border {
@@ -28,8 +54,17 @@
     top: -9px;
 
     border: 1px solid $rose-border-color;
-    // box-sizing: border-box;
-    border-radius: 50px 0px;        
+    border-radius: 50px 0px;
+
+    @media (min-width: 768px) {
+        width: 153px;
+        height: 143px;
+    }
+
+    @media (min-width: 1280px) {
+        width: 153px;
+        height: 143px;
+    }
 }
 
 .img-trumb {
@@ -40,6 +75,16 @@
     border: 1px solid black ;
     border-radius: 50% 0 50% 0;
     overflow: hidden;
+
+    @media (min-width: 768px) {
+        height: 227px;
+        width: 180px;
+    }
+
+    @media (min-width: 1280px) {
+        width: 180px;
+        height: 227px;
+    }
 }
 
 .img-trumb > img {
@@ -56,7 +101,7 @@
     flex-direction: column;
     justify-content: center;
     align-items: center;
-    width: 138px;
+    // width: 138px;
     padding: 10px;
 }
 
@@ -64,25 +109,41 @@
 
 .event-name {
     @extend %event-name-text;
-    display: flex;
-    align-items: center;
-    
     margin-bottom: 5px;
+
+    @media (min-width: 768px) {
+        font-size: 16px;
+        line-height: 20px;  
+    }
 }
 
 .event-data {
     @extend %event-date-text;
-    display: flex;
-   align-items: center;
-
    margin-bottom: 5px;
+
+       @media (min-width: 768px) {
+        font-size: 16px;
+        line-height: 20px;  
+    }
 }
 
 .location {
    @extend %event-place-text;
-   display: flex;
-   align-items: center;
-
    margin-bottom: 5px;
 
+       @media (min-width: 768px) {
+        font-size: 14px;
+        line-height: 17px;    
+    }
+
 }
+
+// .location::before {
+//     content: url(../images/symbol-defs.svg#placeholder);
+// //    background-image: url('../images/symbol-defs.svg#placeholder');
+//     fill: $main-text-color;
+//   display: inline-block;
+//   width: 7px;
+//   height: 10px;
+
+// }
