+++ src/sass/utils/_placeholders.scss
@@ -24,6 +24,7 @@
   font-weight: bold;
   font-size: 14px;
   line-height: 17px;
+  text-align: center;
   color: $hero-text-color;
 }
 
@@ -33,6 +34,7 @@
   font-weight: normal;
   font-size: 14px;
   line-height: 17px;
+  text-align: center;
   color: $main-text-color;
 }
 
@@ -42,6 +44,7 @@
   font-weight: 600;
   font-size: 12px;
   line-height: 15px;
+  text-align: center;
   color:$main-text-color;
 }
